app/backend/requirements.txt
azure-ai-documentintelligence==1.0.0
azure-core==1.32.0
azure-functions==1.21.3
azure-identity==1.19.0
azure-monitor-opentelemetry==1.6.4
azure-search-documents==11.5.2
azure-storage-blob==12.23.1
azurefunctions-extensions-base==1.0.0b2
azurefunctions-extensions-bindings-blob==1.0.0b2
azurefunctions-extensions-http-fastapi==1.0.0b1
python-dotenv==1.0.1
langchain==0.2.17
langchain-text-splitters==0.2.4
openai==1.58.1
opentelemetry.instrumentation.httpx==0.50b0
opentelemetry.instrumentation.openai==0.36.0
